Added the simon-html files
When I cloned it into my repo, it wanted to do so as a submodule. I didn't like that, so I copied the relevant files out of the folder then put them into my repo directly. All files except the .git and .gitignore that was attached to the simon-html have been copied over.